Реализовать авторизацию и регистрацию пользователей и защиту их информации

+++ notes.controller.js
@@ -1,8 +1,8 @@
 const chalk = require('chalk')
 const Note = require('./models/Note')
 
-async function addNote(title) {
-  await Note.create({ title })
+async function addNote(title, owner) {
+  await Note.create({ title, owner })
 
   console.log(chalk.bgGreen('Заметка была добавлена!'))
 }
@@ -12,13 +12,24 @@ async function getNotes() {
   return notes
 }
 
-async function removeNote(id) {
-  await Note.deleteOne({ _id: id })
+async function removeNote(id, owner) {
+  const result = await Note.deleteOne({ _id: id, owner })
+
+  if(result.matchedCount === 0) {
+    console.log(chalk.bgRed('Заметка не была удалена!'))
+    throw new Error('Заметка не была удалена!')
+  }
+
   console.log(chalk.bgRed(`Заметка под номером ${id} была успешно удалена!`))
 }
 
-async function editNote(noteData) {
-  await Note.updateOne({ _id: noteData.id }, { title: noteData.title })
+async function editNote(noteData, owner) {
+  const result = await Note.updateOne({ _id: noteData.id, owner }, { title: noteData.title })
+
+  if(result.matchedCount === 0) {
+    console.log(chalk.bgRed('Заметка не была изменена!'))
+    throw new Error('Заметка не была изменена!')
+  }
 
   console.log(chalk.bgBlue(`Заметка под номером ${noteData.id} была успешно изменена!`))
 }
